Added RatingBar Functionality per row
